19/ena_template_builder/requirements.txt
Flask
Flask-Cors
Flask-WTF
WTForms-JSON
requests
pyyaml
pytest